requirements.txt
lightning >= 2.0.0
matplotlib>=3.0
networkx>=2.0
numpy>=1.20
pandas>=2.0
pymimir
scipy>=1.3
tensordict-nightly>=2024.6.19
torch >= 2.4.0.dev
torch_geometric>=2.0
torchrl-nightly>=2024.6.23
wandb~=0.17.3
